refactor: `int` representations to `sunindextype` where applicable

+++ src/pybammsolvers/idaklu_source/common.cpp
@@ -4,7 +4,7 @@ std::vector<sunrealtype> numpy2sunrealtype(const np_array& input_np) {
   std::vector<sunrealtype> output(input_np.request().size);
 
   auto const inputData = input_np.unchecked<1>();
-  for (int i = 0; i < output.size(); i++) {
+  for (size_t i = 0; i < output.size(); i++) {
     output[i] = inputData[i];
   }
